Create a prelude for the fluent_sdk
